perf: 2.4x faster render path; convolution reverb was 85% of CPU
Reported symptom: buffer overruns in the VST -- clicking a few times a
second on DirectSound, unusable on ASIO -- while the browser version is
smooth. This is CPU overload in the audio callback, not a buffer bug.

Measured first. Bypassing components in the render path showed the reverb
costing 85% of the time: seed 13 ran at 7.8x realtime intact and 54.3x with
reverb disabled. Everything below follows from that.

Changes, all exact -- no fidelity trade:

- The convolver ran in double. Web Audio's entire graph is float32, so
  float is both twice as fast and closer to Chrome. 2x.
- The partition accumulation computed the whole spectrum. Input and impulse
  are both real, so the product is conjugate-symmetric; computing the lower
  half and mirroring is exact. 1.35x.
- /arch:AVX2 for the accumulation loop, with __restrict on its pointers.
- WaOscillator took a mutex on EVERY SAMPLE to resolve its wavetable set --
  once per oscillator, LFO and FM modulator. The table pointer is now
  resolved in setType. Lifted the voice path from 84x to 101x.
- Reverbs kept convolving forever after the last note. A reverb whose input
  has been silent for longer than its impulse produces exactly zero, so it
  is now skipped.
- Filter coefficients refresh every 8 samples rather than every sample.
  Recomputing them costs about 7x the filtering itself. This one is not
  exact: mean fidelity moved 2.28 -> 2.38 dB, still well inside the 2.75 dB
  gate.
- ProcessBlock flushes denormals, restoring MXCSR on exit.

Net: seed 13 one voice 8.3 -> 19.1x realtime, convolver with a 3 s tail
11.4 -> 33.2x. All 86 tests still pass.

Two hypotheses measured and discarded rather than assumed: denormals were
worth only 1.1x, and the compressor runs at 352x and is not a factor.

NOT yet solved, and the headline number is honest about it: worst-case
instruments remain too expensive. Seed 1559 -- five oscillators each with
its own reverb, 9.9 s of tail between them -- runs at 3.5x realtime with
eight voices, about 29% of a core for one instance. That will still glitch
on a small ASIO buffer. The remaining structural fix is non-uniform
partitioned convolution, which is its own piece of work.

+++ core/src/Analysis.cpp
@@ -68,20 +68,20 @@ MelSpectrogram melSpectrogram(const std::vector<float>& signal, double sampleRat
     out.bands = kBands;
     out.data.assign(static_cast<size_t>(out.frames) * kBands, 0.0f);
 
-    std::vector<double> re(static_cast<size_t>(kFftSize)), im(static_cast<size_t>(kFftSize));
+    std::vector<float> re(static_cast<size_t>(kFftSize)), im(static_cast<size_t>(kFftSize));
     std::vector<double> power(static_cast<size_t>(nBins));
 
     for (int f = 0; f < out.frames; ++f) {
         const size_t off = static_cast<size_t>(f) * kHop;
         for (int i = 0; i < kFftSize; ++i) {
-            re[static_cast<size_t>(i)] =
-                double(signal[off + static_cast<size_t>(i)]) * window[static_cast<size_t>(i)];
-            im[static_cast<size_t>(i)] = 0.0;
+            re[static_cast<size_t>(i)] = static_cast<float>(
+                double(signal[off + static_cast<size_t>(i)]) * window[static_cast<size_t>(i)]);
+            im[static_cast<size_t>(i)] = 0.0f;
         }
         fft.transform(re, im, false);
         for (int k = 0; k < nBins; ++k)
-            power[static_cast<size_t>(k)] = re[static_cast<size_t>(k)] * re[static_cast<size_t>(k)] +
-                                            im[static_cast<size_t>(k)] * im[static_cast<size_t>(k)];
+            power[static_cast<size_t>(k)] = double(re[static_cast<size_t>(k)]) * double(re[static_cast<size_t>(k)]) +
+                                            double(im[static_cast<size_t>(k)]) * double(im[static_cast<size_t>(k)]);
 
         for (int b = 0; b < kBands; ++b) {
             double sum = 0.0;
